+++ packages/client/ui/react-native/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @crossmint/client-sdk-react-native-ui
 
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies [89deaad]
+- Updated dependencies [89deaad]
+- Updated dependencies [fcf3c21]
+  - @crossmint/common-sdk-base@0.4.0
+  - @crossmint/client-sdk-react-base@0.2.0
+  - @crossmint/wallets-sdk@0.5.0
+  - @crossmint/client-sdk-base@1.4.17
+
 ## 0.1.1
 
 ### Patch Changes