+++ packages/common/auth/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @crossmint/common-sdk-auth
 
+## 1.0.24
+
+### Patch Changes
+
+- Updated dependencies [89deaad]
+  - @crossmint/common-sdk-base@0.4.0
+  - @crossmint/client-sdk-base@1.4.17
+
 ## 1.0.23
 
 ### Patch Changes